Explain in pxe.md how to change the port on kayak
The kayak webserver by default runs on port 80. Changing this requires
(as far as I know) altering the start/exec line. While there should be
an easier way to do it, until there is, this method should be
documented.

I do take the liberty of pointing out that if you are running a
full-fledged web server on 80, you are better off just using that to
serve the directory.